Project layout altering: moved suggest api package to internal
