+++ internal/suggest/api/reindex_handler.go
